Changed UserAdministration and more
Added a button to UserAdministration to cancel the alteration of a user.
Changed the "wrong input" for adding or altering a user.
Added a minimum character limit to password, the limit is a minimum of 5 characters.
MainMenu now unsets the license data when entering UserAdministration.
Changed IsAdmin to look at the UserID instead of the Username.
Added GetUserName to Connect.php as it is used on multiple pages.
Moved functions to UserAdministrationPHP.
Small changes to the forms.